requirements.txt
absl-py==2.1.0
accelerate==0.27.2
aiohttp==3.9.3
aiosignal==1.3.1
anytree==2.12.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
bert-score==0.3.13
boto3==1.34.48
botocore==1.34.48
bpemb==0.3.4
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
chatoperastore==1.2.0
click==8.0.4
conllu==4.5.3
contourpy==1.2.0
cycler==0.12.1
datasets==2.17.1
Deprecated==1.2.14
dill==0.3.8
docopt==0.6.2
editdistance==0.8.1
filelock==3.13.1
flair==0.13.1
flatbuffers==1.12
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2023.10.0
ftfy==6.1.3
gast==0.4.0
gdown==5.1.0
gensim==4.3.2
google-auth==2.28.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.62.0
h5py==3.10.0
huggingface-cli==0.1
huggingface-hub==0.20.3
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
install==1.3.5
Janome==0.5.0
jieba==0.42.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
langdetect==1.0.9
language-tool-python==2.7.1
lemminflect==0.2.3
libclang==16.0.6
lru-dict==1.3.0
lxml==5.1.0
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.8.3
more-itertools==10.2.0
mpld3==0.5.10
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.2.1
nltk==3.8.1
num2words==0.5.13
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.9.0.80
OpenHowNet==2.0
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.0
pillow==10.2.0
pinyin==0.4.0
pptree==3.1
protobuf==3.19.6
psutil==5.9.8
pyarrow==15.0.0
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycld2==0.41
pygame==2.5.2
pyparsing==3.1.1
pypinyin==0.50.0
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-transformers==1.2.0
pytorch_revgrad==0.2.0
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.10.0
sacremoses==0.1.1
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
segtok==1.5.11
semver==3.0.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.4.0
soupsieve==2.5
sqlitedict==2.1.0
sympy==1.12
synonyms==3.23.5
tabulate==0.9.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.36.0
tensorflow-text==2.9.0
termcolor==2.4.0
terminaltables==3.1.10
textattack==0.3.9
tf-keras==2.15.0
threadpoolctl==3.3.0
tokenizers==0.15.2
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
transformer-smaller-training-vocab==0.3.3
transformers==4.38.1
triton==2.2.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==1.26.18
wcwidth==0.2.13
Werkzeug==3.0.1
Wikipedia-API==0.6.0
word2number==1.1
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
